backend/requirements.txt
fastapi
uvicorn
pymongo
python-jose
passlib[bcrypt]
google-generativeai
pillow
python-multipart